add images for black holes, interviewers
black hole image with spiraling particles. This is only applied to black
holes below a certain size because I didn't think it looked good on big
images. Need to make a better image that looks good for large black
holes.

interviewer is now a face instead of a white box